refactor(direction): ts to sfc
